WikiSearch renamed to SearchResult; int counts -> double values

+++ src/search/RelevanceComparator.java
@@ -2,10 +2,10 @@ package search;
 import java.util.Comparator;
 import java.util.Map.Entry;
 
-public class RelevanceComparator implements Comparator<Entry<String, Integer>> {
+public class RelevanceComparator implements Comparator<Entry<String, Double>> {
 	
 	@Override
-	public int compare(Entry<String, Integer> e1, Entry<String, Integer> e2) {
+	public int compare(Entry<String, Double> e1, Entry<String, Double> e2) {
 		return e1.getValue().compareTo(e2.getValue());
 	}
